lesson_10/requirements.txt
allure-pytest==2.14.2
pytest==8.3.5
selenium==4.31.0
